ld-lights/requirements/base.txt
Django==2.0.1
django-extensions==1.8.1


# PostgresSQL
psycopg2==2.7.3

# serve files
whitenoise==3.3.0

# Rest Framework
djangorestframework==3.7.7

# Sentry
raven==6.1.0

# for date processing
python-dateutil==2.6.1

# Logging
git+https://github.com/Seedstars/django-rest-logger.git

django-rest-knox==3.1.0
